+++ src/Service/Kinesis/src/Result/DescribeStreamOutput.php
@@ -3,6 +3,7 @@
 namespace AsyncAws\Kinesis\Result;
 
 use AsyncAws\Core\Exception\InvalidArgument;
+use AsyncAws\Core\Exception\UnparsableResponse;
 use AsyncAws\Core\Response;
 use AsyncAws\Core\Result;
 use AsyncAws\Kinesis\Enum\EncryptionType;
@@ -174,7 +175,7 @@ class DescribeStreamOutput extends Result implements \IteratorAggregate
             'Shards' => $this->populateResultShardList($json['Shards']),
             'HasMoreShards' => filter_var($json['HasMoreShards'], \FILTER_VALIDATE_BOOLEAN),
             'RetentionPeriodHours' => (int) $json['RetentionPeriodHours'],
-            'StreamCreationTimestamp' => /** @var \DateTimeImmutable $d */ $d = \DateTimeImmutable::createFromFormat('U.u', \sprintf('%.6F', $json['StreamCreationTimestamp'])),
+            'StreamCreationTimestamp' => \DateTimeImmutable::createFromFormat('U.u', \sprintf('%.6F', $json['StreamCreationTimestamp'])) ?: throw new UnparsableResponse('Invalid timestamp received.'),
             'EnhancedMonitoring' => $this->populateResultEnhancedMonitoringList($json['EnhancedMonitoring']),
             'EncryptionType' => isset($json['EncryptionType']) ? (!EncryptionType::exists((string) $json['EncryptionType']) ? EncryptionType::UNKNOWN_TO_SDK : (string) $json['EncryptionType']) : null,
             'KeyId' => isset($json['KeyId']) ? (string) $json['KeyId'] : null,